Move more state to the root component.
Ideally, only the root component has state and all other components just use
props to render. This is accomplished by rendering components like
ApplicationDetail with a null model, and upon mount, trigger the root component
to fetch the model and save it to the root component state, thereby causing a
re-rerender for the ApplicationDetail component with a non-null model.

* Move instance collection to root component state. Instance Details component
  is stateless
* Move volume collection to root component state
* Applications moved to root component state. App detail component uses cached apps
* App detail uses application root state for identities