Removed propTypes hint for `backdrop`.
The Bootstrap Modal component is using `backdrop: 'static'` to affect the behavior of the modal (you cannot click-off on the background to dismiss) and `backdrop: true` to indicate desired styling. So the removing the property type hint seems like the best option to allow `backdrop` to be used in this manner.