Added new field to indicate the preferred UI for a given.
The field `airport_ui` will be a "database null" (e.g. None) as a default.
A truthy value of `not airport_ui` will cause the troposphere UI to be
rendered. It is only when a user _explicitly_ states they want the
previous, old UI that they will be redirected to it.

This also introduces a `?airport_ui=<bool>` query parameter that will
begin to be used to control the rendering user interface (begin with
the K-K, `kicking-kestrel`, release). The `?beta=<bool>` will be
removed in the next release (L-L, yet to be named, release).