Aggressive refactor, cleaner/fixes minor bug
Launch first instance in project, on transition between request-build and
build, the table flashes empty for a sec. This occurs because of some complexity
with tracking pending/regular instances. It happens because of incorrect
filtering behavior in ProjectInstanceStore.

The pending instance pattern is removed. There are some more challenges to
address.

1) ProjectInstanceStore is sort of an alias to InstanceStore, which I made
simpler/more explicit in the ProjectInstanceStore. Having duplicate data in
stores seems bad.

2) the launch action needs further cleanup. There are some assumptions that
could be made more clear. For example, an instance must have a project id.
Also, as it currently stands the instance created (not saved to server), gets
completely replaced from the api, thus the project id for the instance gets
included before/after instance gets replaced.