Altered version view to only provider availability when logged in.
A provider "availability" view needs to know what providers an identity/user
has access to in order to show the "VersionsView". For a public user, so
someone unauthenticated, it is not as important _which_ provider has the
desired version as that "The Cloud" they're looking at *has* it, somewhere.

This is a candidate for refactoring later given how we decide "logged in".
Here, it is just needed _showAvailableOn_ since that relates to the visible
label rendering within the VersionList component.

(cherry picked from commit 3b2c8b23388de22e2e3531db06f3a7b7019f6c9a)