Changed _logged in_ concept to if `profile.get('selected_identity')`.
Determining if a user is logged in has shifted over time from merely doing a
check on `context.profile` to `context.profile.get('username')` to, now, if
a valid `'selected_identity'` is available in the `profile` object.

This notion may **not** hold true across various "clouds", so enriching the
`context` to assign in making these checks (and, really, encapsulating the
details away from the component) is *needed* in the future.