Delayed the fetch of bookmarked images if not logged in.
This implementation is nearly identical commit
21251255f4d89ad82f194c9072bb7885210fae00. Where it departs if the attribute
retrieved from `context.profile` - 'username' in this case. It has usually
been 'selected_identity' within the CyVerse versions of Troposphere UI. We
changed mainly to find a _stable_ indicator of a logged in user being present
within the context (merely have a define `.profile` is not enough after we
changed to `django-iplantauth` module).

Web reference UI for commit:

https://github.com/lenards/troposphere/commit/21251255f4d89ad82f194c9072bb7885210fae00