Use 'username' attribute to determine "logged in" context
A logged in session will always have a value for 'username'.

Why the change? What gives?

Before adopting `iplantauth`, having a 'selected_identity' as an
attribute strongly indicated that the context has a "logged in"
user associated in with session. Now that we are using the auth
module, we can *depend* on 'username', instead. This is true in
all potential authentication configuration for Troposphere (as
of now).