Upgrade requirements to Django 1.8
```
Changed in Django 1.8:
Deletion of the session cookie is a behavior new in Django 1.8.
Previously, the behavior was to regenerate the session key value
that was sent back to the user in the cookie.
```
source: https://docs.djangoproject.com/en/1.8/topics/http/sessions/#django.contrib.sessions.backends.base.SessionBase.flush