Execute expired session clearing on login & logout
Related to triage for [ATMO-1024], the extended session cookie age
duration appears to be masking some of the issues with verification
of tokens grants. Also, the number of _aged_ sessions hanging around
made it seem worth adopting this symmetric approach.

Note - this is just clearing out Sessions. A token is not attached or
linked to a session directly.

(cherry picked from commit 010b71823aa9fe3a6350e53ce1867475a2594cba)