Better validation for ProjectCreateView
There was no visual feedback to validation and trim() was being used on isSubmittable in the test not actually changing the values to be submitted. Since required fields are initialy empty, visual validation would be jarring when the form is first presented. For this reason the visual feedback is only presented after the user attempts to submit with empty required fields.