Provide consistent way to see if user logged in
We are now operating in two contexts:
- there is an authenticated user present
- there is not an authenticated user present

Without logging in, a person should be able to browser the public image
catalog of a given Atmopshere installation (see ATMO-400,1016,1020,
1021).

These leaves components in the dilemna of rendering themselves based on
these two contexts. Previously, components have done checks like:

```
var loggedInUser = !!(context.profile && context.profile.get('username')
```

These are littered around the codebase, and quite fragile. This commit
is the first step in trying to clean this up and eliminate a
delightfully evil "Law of Demeter" violation.