Handle null `end_date` using moment::isValid()
It is possible & likely the an image will have no `end_date`. Changed
the initialization of the endDate variable to look verify isValid()
before populating the input textfield. If the `moment` object refers
to _null_, formatting will result in `"Invalid Date"` - which becomes
visible in the input tag.

Instead of handling this at the Image.js model, it seems best to
leverage the existing method from `moment`, `isValid()`, and check
that instead of doing a null or truthy check.