Bug, missing show more images button
The image select step was only showing 20 images with no show more button. The intent was to compare the total number of images with the current number of images shown and render the show more button if there were more images to show. The issue was that someone had accedentally referenced the same list to compare so the two values would always be the same.