Adopt extract method refactor for (un)auth route handling
Pull out the template parameter population that is nearly the same in
public & auth route handling.

This related to [ATMO-1021] (as well as [ATMO-1020]). The PR for
[ATMO-1021] appears to have been _targeted_ against master, which
would slate the changes for M-M release - not L-L. This work needs
to be included in `larping-loon`. This will be part a PR to get
both public image catalog work (ATMO-1020 & ATMO-400) plus the
default UI (ATMO-1021) in place for release (Feb 23).