Move to *Page.react fetching data
In this `HelpLink`s feature, I'm move to an informal considerations of
the `(MyImages|MyImageRequests|InstanceDetail)Page.react.js`
components as sort of "ControllerViews" (in other words, I want to use
the "block for data & show loading spinner" to ensure that help link
are present before rendering the child elements).

We do not have a statement that a Page.react or View.react is, actually,
a controller view - but I think we likely want to shift the codebase
toward a notion of controller views, presentation views, and _simple_
components held in views. Right now, we have no true guidelines on how
or if or when a child component can _ask_ for something that triggers
a network call (though, I think the original vision for the flux
implementation had that - and maybe it was lost during transition).