Add predicate to indicate auth'd user presence
Previously, we were making various difference boolean expressions that
did one thing - a given a profile model, tell me if there is a logged
in user present (or, put another way, do we have an auth'd _context_).

Since the Header.react component is using a profile passed via `props`,
we don't want to import `context` and say `context.hasLoggedInUser()`.

So, to maintain DRY - this function is added, used by `context`, and
available where a profile can be passed in and the boolean check can
be made.

The implementation is using `!!` to take a 'truthy' expression (the
first `!`) as a conversion, then invert (the second `!`) to Boolean
type representing the overall truthy evaluation. The short-circuit
truthy check avoids errors in the ilk of "undefined does not have
'get' property".

This relates to work associated with ATMO-1279.