Use encapsulate profile check when determining context
When rendering, <Header /> needs to be aware of the auth context. In
other words, it needs to know if it's rendering a public-site view of
the Header or one that has an authenticated user associated with it.

Here, in Header, we use the utility function `hasLoggedInUser` because
the component is getting `profile` passed via `props` and I do not want
to force this component into needed to import `context` when it has the
Backbone model needed to do work (I do understand that this makes it
somewhat of a special case; it - along with Footer).

The meta commentary of the opposite of a component passed profile might
be something like:

...

In essence, we are saying that if you need to get access to the profile
associated with `context` - and you were not passed it via `props`, then
import `context` and call `context.hasLoggedInUser()`.