Use inline styles to recreate .splash-image
When any of the stores hit 503, that's deemed that the API is likely in
maintenance, so if the actual "splash-image" (the loading spinner with
title heading) were shown, then you'd want to removal it from the DOM:

https://github.com/iPlantCollaborativeOpenSource/troposphere/blob/master/troposphere/static/js/bootstrapper.js#L171

However, the className on the div within MaintenanceScreen is not
functionally equivalent because it's holding an image without a spinner
and really should not be re-used.

Given the singular purpose of this component, applying inline styles to
the exact element seems to fit with our project's previous use of
inline style usage.