Handle end-dated instance _sizes_
If you do not hit `api/v2/size?archived=true`, then you will only get
the _active_ sizes (tiny1, tiny2, small1, medium, etc) for a cloud
(by the providers within a cloud).

So, if you are just looking at _active_ sizes, there is a possibility
that an instance's size is no longer available when fetching data to
show statistics or resource consumption, the number of that ilk.

This commit offers two improvements:
- leverage the fact that Atmosphere API does *not* delete things, it
  merely end-dates them
- handle a situation when the 'size' for the Identity model is not
  known (so be robust ... or dare I say _anti-fragile_)

One notion missing from this is the ability to communicate a statistic
is "uncertain" because of the failure to find the associated CPU or
memory for an instance. If `size.get('cpu')` fails because `size` is
false-y (null, undefined, etc) - there is not a way to say "hey, this
CPU count is *kind of* close, but _not_ exactly right".