Remove referencing Backbone.Model's attributes internal structure
We should *never* use `attributes` as it violates the contract Models:

http://backbonejs.org/#Model-attributes

The contract of a Backbone.Model is that you access the properties held
in it via `.get('attr-name')`

Here, I mean for properties & attributes to be synomyms.

You can see the Backbone docs warn about this in relate to setting an
attribute for a Model:

```
Please use set to update the attributes instead of modifying
them directly.
```
source: http://backbonejs.org/#Model-attributes