Use static markup rendering from react-dom/server module
Any of the isomorphic markup rendering that you had once found on React
is likely not *just* within ReactDOM, but on the server variant. I
think this is for server-side situation (hence the isomorphic app
comment).

For all of our charting static markup rendering, you'll want to import
`import ReactDOMServer from 'react-dom/server'`