[fix] Account for undefined `project` prop
In this case, if we are going to use the Object spread operator then
we *must* consider that dot-referring into a prop should account for
undefined/null possibility.

This issue seems to have come about from not knowing the ways a
component may be used. For this, the bug was expose by doing a launch
from the Image Detail page. So using `git grep '<BasicInfoForm'` is one
step - then pulling _up_ from that to where those components used will
help avoid these issues**.

** the other notion we are missing is that we're completely devoid of
   feedback from unit testing or integration testing