Provide an empty state if no status for Volume
There was bug fix, bfe14ddb0d7d9b58becf7ef9e8b21f56a3c24f4f, that
altered the parse for Volume models so that we wouldn't lose info
associated with the Volume (notably `start_date` & `volume_status` -
see ATMO-1116).

The handling of a falsy `attributes.status` meant that attributes.state
was undefined - and that caused a ripple affect in
ResourceStatusSummaryPlot [1] for rendering the state of resources.

[1] https://github.com/iPlantCollaborativeOpenSource/troposphere/blob/c811f88cc3cda9e20ad4d1cc349deb402044367b/troposphere/static/js/components/dashboard/plots/ResourceStatusSummaryPlot.react.js#L56