Ensure boolean evaluates as expected in troposphere.settings
It turns out the `"False"` is a truthy evaluation to `True`; which
implies that we need to be careful with Jinja2 templates - especially
_feature_flags_.

WEB_DESKTOP_INCLUDE_LINK was being evaluated to `True` - which could be
because it was a Python literal boolean, `False`, and a `!!val` coersion
in JavaScript could be `true`.

At any rate, dropping the double quotes around this caused the template,
`troposphere/templates/index.html`, to exclude the `var` from the
globals definition block.

This may be an area we need to investigate slighty more to avoid future
confusion.