Allow a post-create of project to render "Loading..."
There is at least one render cycle where a `project.id` may *not* be
assigned to a Project Backbone model. So this allows the _falsy_ case
to be handled gracefully - in accordance with the rest of the app - by
showing a brief "Loading..." spinner, which will be filled in once the
`done` promise for `project.save()` is reached.