Change UI component to be independent and use indeces
This tab nav now has more internal state. Before it required the parent to set
which tab is the current, now the parent only sets a default.

The other change is that the onChange listener relies an index rather than an
element (see propType defaultLink). There are a few good reasons for this,
it's better to compare against indexes than elements (typically strings).

There were some small name updates.