Initial version of an assets clean-up script
My bash-fu is not strong (though this can be mitigated during PR
review). Here, we have a script that *should* remove all built or
bundled assets from `troposphere/assets/*` - and attempts to
account for OS X versus Linux `find` difference.

The whole `find` _dance_ to surgerically remove files is done with the
hope of not wrecking the directory structure under `assets` (though, it
could be straight nuked outright and `make (dev|cf2|webpack-prod)`
should be able to handle that (perhaps the forceful approach is best).