Define postcss plugin at correct scope [webpack]
So the function that is used as the PostCSS plugin definition was stuck
within `module.exports.resolve` & not at the top level of
module.exports as _required_ by the postcss-loader (major gaff!).

Also, found how to pass in the browser options to `autoprefixer` [1].
This was previously done in the original version using the loader
variation (the new deprecated `autoprefixer-loader`), so it's good that
it is re-introduced here.

[1]: https://github.com/postcss/autoprefixer#webpack