Define variable for WEBPACK_LOADER.CACHE
This will control the frequency that the "webpack bundle" will be read.

**IMPORTANT IMPLICATION**
This means that any uWSGI process will have read the
`webpack-stats.json` file and cached the result. There is no way to
trigger a cache invalidation other than forcing the process to restart
(and, thus, re-read the file).

https://github.com/owais/django-webpack-loader#cache