Create common chunks when bundling JS
This uses as variation from SurviveJS [1] to define a `vendor` chunk
based on all the package.dependencies in package.json. It also enables
webpack.CommonChunksPlugin to create a `manifest` bundle for the app.

During development, I did run into some of the errors described here
[2]. My main issue was understanding that the Webpack loading runtime
needs a function called 'webpackJsonp' - this is part of any chunk
defined as an entry-point [3].

Also, or as a result of the above, the **order** of rendered scripts
_matters_! The `app-[hash].js` bundle depends on `vendor`, it must be
following it then...

So, the correct <script> order was:
- manifest
- vendor
- app / public

[1] http://survivejs.com/webpack/building-with-webpack/splitting-bundles/
[2] https://github.com/webpack/webpack/issues/1315
[3] http://survivejs.com/webpack/advanced-techniques/understanding-chunks/