Change Webpack to produce common asset chunks (#374)
* Upgrade to React 0.13.3

* Add bundle stats tracker plugin (webpack)

* Add hash to assets; restructure PATHS

Bundled assets will have a "hash" component added to the filename to
facilitate cache invalidation when pushing out changes.

The context & output paths within webpack have been restructured to be
part of a JS object in-line with how @bebraw illustrated in the gitbook
"survivejs".

See ATMO-1356.

* [WIP] Initial Django-Webpack loader config

* Define webpack stats target in npm.scripts

* Change to PostCSS loader for styles

* Shift to asset bundle rendering w/ django-webpack loader

This moves us toward loading assets (js & css) with a hash fingerprint
that enables "cache invalidation" when a new version of the asset is
present within the (new) `assets/bundles` directory.

See ATMO-1356

* Ignore bundle tracker file

* Define `assets/bundles` nginx location

* Re-include Webpack Clean for `assets/bundles`

* Initial version of an assets clean-up script

My bash-fu is not strong (though this can be mitigated during PR
review). Here, we have a script that *should* remove all built or
bundled assets from `troposphere/assets/*` - and attempts to
account for OS X versus Linux `find` difference.

The whole `find` _dance_ to surgerically remove files is done with the
hope of not wrecking the directory structure under `assets` (though, it
could be straight nuked outright and `make (dev|cf2|webpack-prod)`
should be able to handle that (perhaps the forceful approach is best).

* Define postcss plugin at correct scope [webpack]

So the function that is used as the PostCSS plugin definition was stuck
within `module.exports.resolve` & not at the top level of
module.exports as _required_ by the postcss-loader (major gaff!).

Also, found how to pass in the browser options to `autoprefixer` [1].
This was previously done in the original version using the loader
variation (the new deprecated `autoprefixer-loader`), so it's good that
it is re-introduced here.

[1]: https://github.com/postcss/autoprefixer#webpack

* Remove cleanup script

Instead, favor the Makefile goal: `make delete-javascript`

* Define variable for WEBPACK_LOADER.CACHE

This will control the frequency that the "webpack bundle" will be read.

**IMPORTANT IMPLICATION**
This means that any uWSGI process will have read the
`webpack-stats.json` file and cached the result. There is no way to
trigger a cache invalidation other than forcing the process to restart
(and, thus, re-read the file).

https://github.com/owais/django-webpack-loader#cache

* Move webpack_loader inclusion to default.py

* Move dev dependency out of dependencies

* Create common chunks when bundling JS

This uses as variation from SurviveJS [1] to define a `vendor` chunk
based on all the package.dependencies in package.json. It also enables
webpack.CommonChunksPlugin to create a `manifest` bundle for the app.

During development, I did run into some of the errors described here
[2]. My main issue was understanding that the Webpack loading runtime
needs a function called 'webpackJsonp' - this is part of any chunk
defined as an entry-point [3].

Also, or as a result of the above, the **order** of rendered scripts
_matters_! The `app-[hash].js` bundle depends on `vendor`, it must be
following it then...

So, the correct <script> order was:
- manifest
- vendor
- app / public

[1] http://survivejs.com/webpack/building-with-webpack/splitting-bundles/
[2] https://github.com/webpack/webpack/issues/1315
[3] http://survivejs.com/webpack/advanced-techniques/understanding-chunks/