Define a utility function to track actions
Using Intercom to note actions, we are looking to capture meaningful
actions in a way to guide usability & user experience design. This
is *not* just a clickstream of "events", but actions that help carry
out the completion of goals and provide an opportunity for support
staff to intervene/assist.

Detailed usage comments for `trackAction` are provided within the
module. Notable, `actionName`s should be a past tense verb joined
together with a noun by a delimiter:

- 'created-project', 'reported-instance', 'sent-feedback'

These actions can be used to segment and define a community cohort.