Improve eslint (#388)
* Add travis lint as a special case

Include .eslintignore to make eslint commands shorter

* Make .eslintrc effective

I ran eslint with the recommend rules which is pretty strict. Then I gathered
the list of rules that troposphere was breaking. I separated the serious rules
from the innocuous ones. .eslintrc ignores the silly ones but still respects
the serious ones.

* Add documentation for linting

I added some of my reasoning in separating the various .eslintrc's and added a
howto for running the linter.