Altered render for instance row
When an instance does **not** have a valid database issued ID, then
avoid rendering with the temporary "child ID" (aka, cid). Also, if
no `instance.get("id")` is present - disable the Instance Delete Modal
button for deleting.

Ideally, the InstanceDeleteModal would render a helpful indicator for
the state the user interface is in. Or, better yet, the SelectableRow
would not **be** _selectable_ when there is no `instance.get("id")`.

The root of this appears to be that `instance.id` is returning a "cid"
when it previously was returning `undefined`.