Discontinued combined cert creation from Makefile
This actions will be moved into Clank and be done on a *conditional*
basis, depending on the SSL configuration desired by those running
Clank.

It is believed to be the case that a "combined" may not be used in all
situations. Given that CyVerse organization has a wildcard certificate
bundle - we can, and do, operate using this.

If you are using a certificate generated from an authority, the
combined certificate appears to be similar to a "fullchain.pem".

Thank you to "Let's Encrypt" and EFF for their efforts in making SSL
certificates available.