Removed "new window" logic from Web Desktop link
This is only a temporary fix, but the action-link for "Web Desktop"
had `openInNewWindow` and cause it to be *disabled* due to a lack
of a `href` reference - however, we do now truly _know_ the link
until we defer to `views/web_desktop.py`.

We *could* put in a `href` value for `views/web_desktop.py` and
allow for the onClick to override it.

For this temp fix - I've opted to just omit the "external link"
styling.

(cherry picked from commit f3d64315abff0f035594b98786c56c40bda36d5d)