Allow SelectMenu to handle a null option
Here is the original interface of `SelectMenu`:
<SelectMenu
    current={ The element to display }
    optionName = { A function mapping an element to a name }
    list={ The list of elements }
    onSelect={ The function to call on element selection } \>

It allowed list to be null, which it interpretted as loading. But it didn't
handle a null current, which is a common use case in select menus (i.e.
nothing is selected). The new interface addresses this issue.

<SelectMenu
    current={ null }
    placeholder={ Text to display if current is null }
    optionName={ ... }
    list={ ... }
    onSelect={ ... } \>

Now a select can have an empty selection, and display optional placeholder
text.