[fix] Ensured modal knows the project.id of selection
The selection shown in <ProjectSelect /> is not communicating the
displayed options back to the modal. Told from a different view,
the modal thinks that the defaultValue for the `<select ../>` is
this.state.projectId = -999 because `getState()` was not being
called prior to render.

Therefore, the only way that this.state.projectId would be set
correctly is if an action was taken to trigger `onChange`.

This fix added a called to `this.updateState()` to trigger the
code used to set the defaultValue for `<select ../>` to be the
first element in the project collection - if available. When there
are no projects, the sentinel value `-1` is used to force a project
to be created.

This component should be re-evaluated in the future. Even if we remove
the "project-less" user interface (aka airport-ui), there are *still*
situations with cloud providers that can create this scenario (e.g.
resources without an assigned project).

Also - we should consider doing a search of the database to look for
users with a preference of "Troposphere" & an instances/volumes where
`projects == []` as it takes a *new* project-less resource to trigger
the NullProjectMigrateResourceModal to be shown.