Altered if Delete resource button available
This work removes the "Delete" resource button multiple resources are
selected in the ProjectDetail view.

Background:
-----------
Currently, we do *not* support a smooth "batch" delete interaction.
Prior to this commit, we deleted the _last_ selected resource. In
an effort to avoid confusing affordances, this is no longer visible
while buttons that can be used in "multi-select" remain (notable here
is "Move selected resources" - which is why the "checkbox"/multi-select
remains in the UI at all; otherwise, a simplification would be to pull
the checkboxes out altogether).

Future Work:
------------
The move to a MediaCard interface will present the opportunity to
address the topic of "Batch Operations" again. It is important that
we offer this is a manner that makes organization within Projects
beneficial.

See JIRA tickets: ATMO-981, ATMO-1467