Upgraded from Babel 5 to 6
Shift to "presets", we use ES2015 features along with React & Stage-2
functionality (notably the spread operator from stage-2). This means
the previous `.babelrc` properties "stage" and "optional" were
replaced in favor of "grouping" them as _presets_. [0]

We also include `babel-plugin-add-module-exports` because Babel 6 does
do a default `module.exports` (which we want). This is a plugin and
included in `.babelrc` [1]

Thanks to Serg Gospodarets for his post [2] and to James K Nelson [3].

[0] https://babeljs.io/blog/2015/10/31/setting-up-babel-6
[1] https://www.npmjs.com/package/babel-plugin-add-module-exports
[2] https://medium.com/@malyw/how-to-update-babel-5-x-6-x-d828c230ec53
[3] http://jamesknelson.com/using-es6-in-the-browser-with-babel-6-and-webpack/