Add fallback to find peer dependencies
This fixed an issue where webpack could not resolve an npm linked dependency's
peer dependency.

Specifically webpack could not locate react in the tropopshere-ui symlink. It
was looking for react in troposphere-ui.

Suppose the following installation:
/opt/dev/troposphere
/opt/dev/troposphere-ui

When webpack resolves a module it looks in the modules node_modules, then the
parent node_modules and up. However, in the case of troposphere-ui, webpack
should look for react in /opt/dev/troposphere/node_modules, however this is
/not/ a parent of troposphere-ui so it's never found.