Avoid case where findWhere does lookup on `undefined`
For the situation where a response from v2/resource_requests returns
a response of `[]`, the resulting state of ResourceRequestStore is a
collection contain 1 element (a Model) with no attributes.

```
  var lookup = model.get(tokens[0]);
  if (lookup[tokens[1]] !== params[key]) matches = false;
```

This is far too optimistic, but since we're dealing with a response
from an internal API - I can see where it may have been assumed that
the response was going to have "some" models for inspection.

So, here we could have perhaps said:
```
if (!(lookup && lookup[tokens[1]] == params[key])) matches = false;
```

This gets us the "short circuit" evaluation that avoid looking for
`tokens[1]` in _undefined_. But to keep with the "reverse", or
negation, logic - I've just applied a De Morgan's Law conversation
to the expression:

!(A && B) becomes !A || !B