Conditionally add Array methods for older browsers
Two methods on `Array.prototype` appear to be missing in some browsers.

Thanks to Sentry & raven-js, we know that some browsers are getting
Uncaught Errors for `Array#includes` [0] and `Array#findIndex` [1].

We would expect that a browser missing the ES6 (2015) `findIndex` would
be overlayed by a babel-polyfill. This is still a bit unexpected.
However, the `includes` method is defined in ES7 (2016) and it was not
a _reasonable_ assumption by me (@lenards) to allow code in without
providing an appropriate polyfill to catch browsers without it.

Again - thank you to Sentry for the insights.

[0] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/includes#Polyfill
[1] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/findIndex#Polyfill