Implement "export" of credentials for client tools
This implements an advanced settings option to "Export Credentials" on
a _per Atmosphere Provider_ basis to be used with CLI tools for the
OpenStack APIs.

The credentials are provided for "Copy & Paste" into a shell script file
that will _export_ the environment variables used by tools like
`openstack` when sourced. [1]

This is a basic, "Phase 0", version of the feature.

See ATMO-1457

[1] https://pypi.python.org/pypi/python-openstackclient