Adjust min-width to ensure that *hash* OS_PASSWORD fits
The visible <pre> for the export credentials was not seen during dev
because I was using `<REDACTED>` as the value for OS_PASSWORD.

The length of a generated value for OS_PASSWORD is 64 characters. When
you include the environment variable and `export` keyword, the line is
83 characters (shown below with password placeholder):

```
export OS_PASSWORD=****************************************************************
```

A minimum width of 705px appears to be the minimal pixel count on my
display to get the value in full view for copy and paste. Even if the
value occluded, the <pre> will scroll and can be selected for
copy-paste. The *reason* I have went to this trouble is because
OS_PASSWORD is the last value within the monospace block and it may
not be 100% obvious is a text-selection was such that it do **not**
include OS_PASSWORD

This will be mitigated by a "copy to clipboard" usage.