Ensure required field has instanceName value
This has the potential to be slightly frustrating to a community member.
Let us just state that right away.

The situation is this, someone deletes "Instance Name", tabs out of the
field and moves on to other things. When they try to launch, it will
*fail*. So, why not be a guardian angel and just _prevent_ them from
entering into a situation that will lead to a failure?

That is the notion here.

If we have a `image` model, then get the name and fill it in for
`instanceName`. If that is not available, just use some placeholder text
that would be easy for a Cloud Operator to identify as having come from
Troposphere.

See ATMO-1595