Add "expired" password badge & info to context menu
One of three scenarios for bubbling up the presence of an expired
"delegation" password is the "LogoutLink" shown within the Header.

The component <LogoutLink/> displays the `username` of the community
member with a caret to denote a "context menu" hidden within it. This
is shown `onClick`

So here, we mark the visible `username` with a badge (!) to the left
of the name and *insert* into the context menu as notice of the issue.

See ATMO-1481