Add "expired" password badge & info to launch button
One of three scenarios for bubbling up the presence of an expired
"delegation" password is the "Launch" button shown within the Image
Catalog (for only an authenticated user).

The component <HeaderView /> displays the image's information and
exposes the ability to "Launch" from this catalog-detail-view. Here,
we add a "badge", (!), to the upper-left of the button and replace the
`onClick` with a "show" modal explaining the issue.

See ATMO-1481