Use W3C Clipboard API for copying credentials
Information about the W3C Clipboard API [1] makes it possible to copy
and paste DOM nodes (aka elements) without using Flash or an external
library.

Steve originally discovered this on SO [0] (huge thank you to Dean
Taylor [*] for his detailed steps to utilizing the APIs).

According to the MDN entry, browser compatibility [3] for the copy/cut
API is pretty good:

 - IE9+, Google Chrome 43+, Mozilla Firefox 41+, Safari 10+, Opera 29+

What we're doing here is providing a utility module to allow use to
check for the functionality [2] and then making use of it.

The examples originally reviewed [0] used "input" elements, which have a
`select()` function [4] for performing the range selection. The code
added in this commit is allows for _general_ element copying by using
Range and Selection objects from `document`.

Note: we clear all "active" selections here with `removeAllRanges()` to
avoid a "discontiguous selection is not supported" [5], thank you to
the authors of clipboard.js and copy-to-clipboard for all their efforts.

[0] http://stackoverflow.com/questions/400212/how-do-i-copy-to-the-clipboard-in-javascript/30810322#30810322
[1] https://w3c.github.io/clipboard-apis/
[2] https://developer.mozilla.org/en-US/docs/Web/API/Document/execCommand
[3] https://developer.mozilla.org/en-US/docs/Web/API/Document/execCommand#Browser_compatibility
[4] https://developer.mozilla.org/en-US/docs/Web/API/HTMLInputElement/select
[5] https://github.com/zenorocha/clipboard.js/issues/17

[*] http://stackoverflow.com/users/406712/dean-taylor