Specify <Router /> with browser history
This introduces a "wrapping" function to set the basename name for
browser history to be `/application` using the `history` library's
`withBasename` function.

This is an approach widely used through the examples wihtin react-router
repository [0]. This will apply `/application` as the basename when
creating browserHistory.

NOTE: it seems that `withBasename` is removed in version of `history`
beyond version 2.1.2 (which is the version paired w/ v2.8.1).

[0] https://github.com/ReactTraining/react-router/blob/v2.8.1/examples/withExampleBasename.js