Resolve ImageVersions not being loaded
This is a provisional solution that needs to be addressed in the future

... it seems that with the React 15.* & react-router, we are seeing that
the "bulk" event listener *adds* in Master.componentDidMount do _not_
appear to be triggering the "renders" within the "children" included for
*that* route (errr, "route component" - maybe).

So we have this situation where we desire an "updateState" call on the
composite (user-defined) React Class, and no real _workable_ to "signal"
the need - other than a faux `this.setState({signal:"yay"})` or a call
to `this.forceUpdate()` (a known anti-pattern).

This is moderately troubling - I resolve to be wiser RE this soon.