Migrate Route definitions to v2.8.1
The naming of attributes used within a `<Route .../>` changed in v1 [0]

This effort changes all AppRoute definitions to comply with the new
format overall.

One thing to note is that participating in a route makes a React Class
a "Route Component".  This is helpful to know when reading the
react-router documentation and their upgrade guides. A "Route Component"
will have inject objects/values on `this.props` which are *not* present
for a class not participating in a route [1,2]. This changes the
approach to making available browser history or router state, etc.

(This is more a message to "Future @lenards" - but anyone else that
finds it helpful it _great_ too)

[0] https://github.com/ReactTraining/react-router/blob/master/upgrade-guides/v1.0.0.md#route-config
[1] https://github.com/ReactTraining/react-router/blob/master/upgrade-guides/v2.0.0.md#goals
    > Draw a clean line between what goes to Route Components as props
    > and what goes on context.
[2] https://github.com/ReactTraining/react-router/blob/master/upgrade-guides/v2.0.0.md#navigating-in-route-components
    See section anchored to above versus "Navigating inside deeply
    nested components"