Fix the data fetching
Previously render only used get store methods, which never fetch and only read
if data has been fetched. This meant that prop changes (a new query) didn't
always cause data to be fetched.

Now when render gets called the data gets fetched if it's not there. Render
always gets called on mount or when a prop changes. This is simpler than
knowing that when a prop changes updateState gets called, and updateState
makes a network fetch.