Avoid a leading & trailing newline added to text
When `display: table` style is applied to <Code />, a leading newline &
trailing newline are added to the text copied to the clipboard. This
causes pasting into a command-line interface to be problematic as the
argument is not applied correctly.

```
nova list <pasted>
```

 turns into
```
nova list # command runs, and fails
d559c236-53ef-437a-999e-02ae16c4b5f0 # command runs, and fails
command not found
```

With this change, we do not add any whitespace.

This problem space was explored in a codepen:
http://codepen.io/lenards/pen/rjVwLP

Thank you to @Tharon-C for his expertise with CSS and immediately
knowing the solution to this. /bow