Allow *bulk* suspend actions for instances
- Restores "suspend" on multiple instances

This allows for a user-directed action of "suspend" performed on all
selected instances within the Project-Details view.

Previous, "selected" actions would only be performed on the *last*
resource selected and *not* all resources. An action was taken to
disable all "bulk" actions for selected resources to attempt to hide
the "inconsistent" _perform-action-on-selected-resource_.

This is outlined internally in the ATMO-981 ticket. Overall, a member
of the community noted that the behavior was highly confusing. But the
move to "disable" remained in the user interface unaddressed for too
long, which I (@lenards) take full responsibility for.

This goal in this and forthcoming commits to allow:
- suspend all selected instances
- stop all selected instances
- resume all selected instances

The Delete action will be handled in a seperate pull request.

And, I (@lenards) apologize to each click-wary community member for
having to select-and-suspend/stop one-by-one.

See ATMO-981 [0]

[0] https://pods.iplantcollaborative.org/jira/browse/ATMO-981