Pluralize resource handling in instance modals
Previously, the Instance*Modal components were only handling actions on
a 'single' instance - not many.

Also, modals did not incorporate information about the instance
affected if they were to _continue_. The `props` passed to modals was
almost always `null`, so there was not a chance include the models for
improved "context".

The changes in the work associated with ATMO-981 [0] will allow for
actions on "batches" of instances.

NOTE: "selected resources" could include models of type:
- Instance
- Volume
- Image
- ExternalLink

The potential for heterogeneity in resources leads this modals to do
"filter" operations via `instanceof` to ensure they're giving correct
feedback to the community member about *what* instance will be acted
upon if they click "Yes".

[0] https://pods.iplantcollaborative.org/jira/browse/ATMO-981